refactor: api server using best practices
