+++ api/common/constants.js
@@ -3,5 +3,6 @@ dotenv.config();
 
 const JWT_SECRET = process.env.JWT_SECRET;
 const AUTH_TOKEN = "auth_token";
+const UPLOADS_PATH = "/uploads";
 
-export { JWT_SECRET, AUTH_TOKEN };
+export { JWT_SECRET, AUTH_TOKEN, UPLOADS_PATH };
