+++ api/models/Post.js
@@ -1,4 +1,4 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
 const { Schema, model } = mongoose;
 
 const PostSchema = new Schema(
@@ -17,4 +17,4 @@ const PostSchema = new Schema(
 
 const PostModel = model("Post", PostSchema);
 
-module.exports = PostModel;
+export default PostModel;
