+++ api/models/User.js
@@ -1,4 +1,4 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
 const { Schema, model } = mongoose;
 
 const UserSchema = new Schema({
@@ -9,4 +9,4 @@ const UserSchema = new Schema({
 
 const UserModel = model("User", UserSchema);
 
-module.exports = UserModel;
+export default UserModel;
