--- api/paths.js
@@ -1,15 +1,0 @@
-const PATHS = {
-  UPLOADS_PATH: "/uploads",
-  LOGIN_PATH: "/login",
-  LOGOUT_PATH: "/logout",
-  REGISTER_PATH: "/register",
-  PROFILE_PATH: "/profile",
-  CREATE_POST_PATH: "/posts/create",
-  ALL_POSTS_PATH: "/posts",
-  GET_POST_PATH: "/posts/:id",
-  UPDATE_POST_PATH: "/posts/:id",
-  FEATURED_POST_PATH: "/posts/:id/featured",
-  DELETE_POST_PATH: "/posts/:id/delete",
-};
-
-module.exports = PATHS;
